requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
django-filter>=24.2
pytest>=8.0
pytest-django>=4.8
